Add lots of new Sokoban levels from Slex
There are now 57 varieties of Sokoban.
The levels were ported from Slex's sokoban.des with major changes,
including new zoo selection, new loot selection and deduplication
(replacing multiple copies of the same map with one copy using
randomization). The overall power/deadliness level has been much
reduced, but it should still be more of a challenge than before.
In the process a new room type (lab) has been added to the level
compiler, and the dat Makefile now preprocesses the .des to allow
macros instead of copy-paste. (The built-in functions unfortunately
aren't up to it.) This will only work on Unix so both input and the
resulting .des are shipped.