Add magic chests from dNethack
Magic chests, ported from dNethack. This is a container with shared
contents - items put into a magic chest can be retrieved from any
other magic chest. There are 10 compartments which can be locked
separately. The chest can be moved, but it's very heavy - this weight
is constant though, independent of the weight of the contents.
They are randomly generated, rarely.