Add climbing boots ("shiny boots")
Add the Climbing attribute, and climbing boots which are the only
source of it. These have the following abilities:

Can climb stairs when Strained or less.
Can descend stairs when Strained or less without falling down.
Will always succeed at getting out of a pit.
Can safely climb down into a pit, or through a trapdoor or hole,
	if the trap is known (and you are not riding or blind, etc.)
Can squeeze past (over) a boulder with more inventory.

Side-effects of this are:

Trapdoors and holes will now do a small amount of damage, like pits.
The test for squeezing past a boulder has been changed from "at least
850 aums less than your maximum capacity" to "no more than 75 aums".
(Or 350 aums if you have Climbing.)
(This is more reasonable as this is limited by bulk - it's not a test
of strength.)