Make magical explosions more merciful
When a bag of holding goes BOOM, most of the contents are thrown
around the area (within LOS). Blessed items and (especially) artifacts
are less likely to vanish, cursed ones more so and Luck makes a lot of
difference. Surviving items get dropped nearby - with or without impact
(destroying fragile items, as if dropped while levitating) depending
on the same chance as vanishing. (Items that were flung further are
also more likely to shatter.)