Add sub-roles.
Sub-roles are a way to define a subset of a role which can be selected
at character creation. They take effect by changing the restrictions
(alignment, race, gender) on the role, by changing members of urole and
by containing a subrole ID which can be checked with Subrole_if and
Subrole_switch, which behave like Role_if and Role_switch. They are
intended for variations on a role which don't require a whole new role,
especially if they are to share the same Quest.

This patch includes an empty subrole array (and example template) and so
should not change behavior by itself.

This patch breaks savefile compatibility and requires window system
support - it includes support only for curses.