Support an internal RNG
Adds support for an internal RNG, i.e. redirection of all calls to
seed or generate random numbers to new functions seed_rng() and
get_rng(). These are in a new file rndint.c, so the Makefile has been
updated - as yet only for Unix.

This patch does not add an implementation of these functions, so
*it will break* unless you add a replacement.