Amulet versus stone
It provides petrification resistance, as you would expect.
It also stops stoning when you put it on - but only if not cursed, and
losing one point of BUC in the process.
It also deflects thrown rocks and really doesn't like stone golems,
giving much increased damage against them. (But DON'T wear it if you
are polymorphed into a stone golem!). It also speeds digging.