Alchemist sub-role of Wizard
The Alchemist is a variant Wizard with starting equipment biased
towards potions (especially potions of acid), along with an
alchemist's smock.

Alchemists can brew potions using recipes which involves dipping
one or more items into a potion of acid. They blow stuff up a lot
but can create any potion - limited by the availability of the
ingredients and the new Alchemy skill. It is intended to complement,
not replace, the existing potion-mixing effects which have not been
modified.

This has not had much testing yet and probably needs some tweaking.