Zombie Apocalypse Patch
Zombies are no longer boring.

They have a attack which can (with chance increasing for higher level
zombies) drain one or more points of INT. This can't kill though
brainlessness or cause memory loss - it's not the same as a mind
flayer's attack. (This is not expected to make things much more
difficult - they are slow and can be attacked from range, and you
won't have encountered many before you get a a unicorn horn.) It can
also cause confusion. (Anything describes as a zombie has this attack,
as do ghouls. Draugr and skeletal pirates are Z's but are weapon users
so do not. Nor do skeletons - they are weapon / slow-touch users.)

If a zombie kills a monster with a zombie form, it will rise again as
a zombie. Liches also have this ability, and zombies will attack any
monster which can be zombified. If you are killed by a zombie, the
same thing happens and there will be a zombie instead of a ghost in
the bones file.

A few new zombie types have also been added, to provide more targets
for zombies to attack (and add some variety).