Pet tweaks
The rate of wildening when you are off level is unchanged on average,
but is randomised to make it less predictable.

Pets are less likely to kill themselves. The existing behavior of
not attacking when below 25% HP remains. When above 80% max HP the
behavior is unchanged - they will attack anything up to 1 level
above them. From 60 to 80% they will attack up to their own level
only, from 40 to 60% up to 1 level below and from 25 to 40% up to
2 levels below.