Alternative rne for enchantment/charge
Replace random enchantment/charge of weapons/armor/rings with a
function which is distributed exponentially, like rne(), but
biased by Luck rather than level. It uses chances based on the
probabilities of scrolls of enchant armor, enchant weapon and
charging blowing up an overenchanted item, or having no effect,
so while it is open-ended for weapons and armor, enchantments
higher than the highest you could safely reach are rare. Set
against that, you don't have to wait until level 18 - by which
time it is unlikely that you would care. Rings are not open-ended,
because a ring higher than +6 will always explode when charged.
Charging can be by up to 3 points, so +9 is the highest you can get.