Praying for food is smarter
The previous behaviour was to treat any prayer while Weak or worse
as a serious trouble. It will now first check your inventory for
food (and your bags, if they are accessible to you) and if there is
a significant amount you will be expected to eat that first.

Many exceptions are made - for tins (because of the delay), tripe
(which might make you sick), eggs (which might be cockatrice), and
items which are not intended as food (catnip, garlic, wolfsbane,
eucalyptus, lizards). Corpses and globs are only considered if they
won't do anything too nasty (stone, slime, poison, acid if low HP,
domestic animal, cannibalism, or tainted).