Fix weights for items with changed materials
In several places an item was being generated with a random material
and then the material was set by changing the object's material
directly, without changing the weight. Usually this is unwanted,
producing items with a different weight than an item of that material
would be expected to have. Change these places to use the new function
set_material() which sets the weight and erosion/erodeprooofing - and
in future could set other material related properties.