Bring back the scrolls removed by the Splice 0.6.0 merge.
Splice 0.6.0 removes the scrolls of pure law, true neutrality and
raw chaos. This adds a similar replacement - the scroll of divinity.
This doesn't care about alignment (because otherwise 2/3 of them
are junk, and such a rare scroll should not be junk) and buffs them
by allowing them to work in all cases, even when the prayer timeout
was preventing prayer.

Blessed scrolls also reduce the timeout after prayer by up to half,
depending on luck, while cursed scrolls cause a failed prayer - with
possible divine retribution.

Reading it while confused has an effect which is probably not what
you wanted :)