New item - gauntlets
Gauntlets are non-magical iron gloves with 2 base AC.
Some knights will start with them (and no helmet - so the same AC.
The helmet is more useful against traps, etc. but easier to find so
it doesn't make much difference to difficulty.)
They are as rare as the magical gloves in random generation, but the
watch and soldiers may have them.