Add back tiger-vs-tiger eye ring
This was removed in the merge.
This version though, for consistency with the new silver ring code,
does damage on half of hits if you have one ring and on all hits if
you have two rings (rather than on all hits, doubled if you had two
rings as it did before).