Object probability encoding
Object class probabilities of 1000 or less behave as before - that is,
the chance to generate that object (compared to the total of all
objects in that class, which may be but is not necessarily 1000).
So this should not break compatibility with existing objects.

Values of oc_prob above 1000 describe a minimum level to generate the
object (value / 1000) and a chance to generate the object (value %
1000). Lower depths give a probability of 0, similarly to gems.

This chance is scaled down by a factor of 10 compared to standard
(<= 1000) generation, so this encoding can also be used for more
precise control of rarity (especially for very rare items).