Add a new item, the holy symbol.
The holy symbol is an amulet which when worn tells you whether it is
safe to pray - it's blessed if safe to pray, uncursed if the timeout
prevents prayer, and cursed if there is some other reason. (It also
gleams, blackens, etc. when changing BUC.) Because of this, you can't
curse or bless it while it is worn. Trying to do so will make a small
difference to prayer timeout, though.

Wearing it also speeds the prayer timeout by 25% - or 50% if you are
a Priest - and if worn while praying successfully you may get a better
effect than you otherwise would have. This does depend on Luck.

If uncursed, it can also be invoked for turn undead. This has a long
timeout, shared between all holy symbols (so that getting lots won't
help you). Invoking it when it is cursed instead summons undead!

If sacrificed, it is a high value similar to a cross-aligned unicorn.
Also like a unicorn, sacrificing it on a cross-aligned altar will
convert you. Wearing it while sacrificing a corpse will increase the
value similarly to, and cumulative with, using a sacrificial knife.

If dropped, it will scare vampires (like an altar). Wearing one is OK
for atheists (if fairly useless), but turning undead is not.

They are uncommon, but priests and similar monsters may carry one and
there is a small chance of finding one as the Sokoban reward.