New Dragon Armor
This patch introduces dragon shields, dragon helms, dragonhide
gauntlets, and dragonhide boots. These items have the same extrinsic
that the mail and scales provide and 3 AC. To make them you need to
wear scales in the shield, helm, gloves or boots slot and read a
scroll of enchant armor. This means that this patch also incudes
support for multislot armors, and because this makes the unenchanted
scales a more powerful item they are now a little heavier and have
only 1 AC.

A confused scroll of enchant armor's usual use doesn't apply to dragon
armor, so it can now be used to turn scales into a random dragon armor
(this is not useful) or turn a dragon armor back into scales (which
might be useful, if you wanted it in a different form.)

These are powerful items, and the intent is to give more variety to
ascension kits but without making things overly powerful and trying to
avoid items being strictly better than an existing item - although not
if this is at the expense of consistency.

For this reason, dragons are classified as "common" (chromatic),
"rare" (black, silver, grey, shimmering, purple), "high" (hex, filth,
ooze, razor) and "never generated" (void). You can wear common items
in any combination, but you can only wear one colour of rare (or
better) items at any time, and so only get one property from them.
(So, for example, that silver dragon shield is lighter and more
protective than a shield of reflection - but it won't cooperate with
your grey dragon scale mail.)

All but the void dragon scales are randomly generated, but only deep
and very rarely. You are still expected to be more likely to get
dragon armor by killing a dragon - or wishing for it. They are also
now considered magical, so you can't easily polypile for them.