Theme chests
Some chests and large boxes (more often those with a lot of stuff)
are 'themed' which at the moment just means that all items are from
the same class. The numbers of items in a box has also been changed,
so than occasionally larger numbers of items will be seen, and locked
and especially trapped boxes usually have some loot in. (The average
number of items generated in containers should not have changed much
though, and the distribution of item classes should also be
unchanged.)