HP change - split Con from level
This change splits the part of level-based max HP gain which is
based on Con into an independent bonus which is updated whenever Con
or level changes. The result is that your current max HP depends on
your current Con, rather than the history of Con at level gain.

As most characters gain Con over the game, the average Con gain has
been adjusted down slightly (in the previous patch) intending to
end up with no change to typical HP. Other effects are that items
which boost Con are more valuable, while loss of Con is more
immediately dangerous (although without any long-term effects once
restored). It is possible to lose 1HP on level up, but only if you
have 5 or less Con. Some kinds of "drain for gain" are prevented.