Add the artifact "Quick Blade".
Quick Blade is an (iron) saber with the property that a successful hit
with it takes half the usual turn time. The result is comparable to
double damage but is more powerful (attacks can be split between
opponents, especially when one dies with the first hit) and reliable
(with two attacks against the same opponent you have another chance
to hit).

That does only apply in the ideal case of hitting every time, though.

You can expect 9 average damage at +0 and 23 at +7 - the same as
Grayswandir, but without the silver damage.

It's intended to be a AK-worthy but not overpowered (compared to Frost
Brand, for example) weapon that encourages a different tactical style,
with Archeologists (as the only role which can gain Expert in saber)
being especially targeted... though it probably needs some tweaking as
yet.