Add a modified version of Topi Linkala's Owned Artifacts patch
Artifacts have an owner. If you wish for an artifact there is an
increasing chance with each wish that you will get the artifact
only after killing its guardian. These are generally player-monsters
of a type associated with that artifact (e.g. blunt instruments get
a Priest, Stormy gets a Barbarian...). There are a few special cases
(Thiefbane, King in Yellow, Vladbane), and quest artifacts will
sometimes bring in the associated nemesis!