Port Stevie-O's level teleporter jump patch
^T while standing on a level teleporter activates the trap.
Sometimes (especially if you have teleport control) you *want* to use
a teleportation trap. Pressing ^T while standing on a intralevel
teleporter will print "You jump onto the teleportation trap." and
activate the trap. For a level teleporter, however, the only way to
use the trap is to remove any and all objects providing MR (CoMR, Eye
of Aethiopica, Grey DSM, wielded Magicbane, ...), step off of the
trap, and then step back onto it.

This patch modifies the ^T command slightly so that if invoked while
standing on a level teleporter, the game asks "There is a level
teleporter here. Jump in? [yn]". If the user chooses 'y', the
teleporter is activated as if you had no MR. Choosing 'n' will
fall through to the original logic (using the Teleport Away spell,
or exploiting teleportitis with TC).