Patric Mueller's Harmonic Scoring
Limits scoring by changing the score per monster kill.
Scores grow without limit. For every killed monster the player gets
the same number of points.

For regular monsters this is not really a problem as after a specific
number of kills these monsters won't be generated again. Unique
monsters like the Wizard of Yendor or the riders can be killed
infinite times.

The Harmonic Scoring Patch decouples the score points per monster
kill from the experience points and decreases them continuously with
every kill.

For the first kill, the complete points are awarded, the second kill
yields only half points, the third kill only a third, etc.

This leads to smaller scores. For the early and mid game scores are
about half as large. For the late game and pre-ascension, the score
is between 0.5 and 2 million points.

With this method of scoring the doubling of the score and counting of
the value of the items for an ascension have again a noticeable impact
on the final score.