Castle soldiers don't fall through the trapdoors
It is common for lots of soldiers and other Castle monsters to fall
through the trapdoors into the Valley. This doesn't make much sense,
they set the traps after all and should know where they are - and it
doesn't help the appearance of the Valley either.

So monsters will now avoid a trap (as if they had already encountered
one of its type before) if:

It is the Castle, Ludios or Quest End level (it's not an obvious
problem for Ludios or Quest End, but the same reasoning applies),
and the monster is a non-mindless humanoid.