Fix a couple of Alchemist bugs
The "choose subrole" prompt is removed - it never worked, and it is
probably a bad idea to try to keep it as a full list of subroles may
some day exceed the screen's height.
Also change the way rarities are calculated - in some circumstances
this could crash at character creation for an Alchemist.