Improved Wish
An improved wish is one in which specified numbers (enchantment,
charges, size of stack) are biased to allow you to get more (or
less, although that's not used yet).

The only way to get an improved wish is to break a wand of wishing
which either has at least one charge left, or which has not been
recharged. (Not having been recharged is equivalent to one more
charge). Such a wand can (with the wrest charge) be used to make
at least two wishes, which you are giving up for one more powerful
wish.

Each additional charge increases stack sizes by 50%, maximum wand
charge by 1 and maximum armor or weapon enchantment by 1. It doesn't
otherwise change the way that enchantment is handled, so it's still
not guaranteed to get more than +1 - although the average is higher
and it is possible (if very chancy) to get more than +5.