Add Wizardstaff
A wizardstaff is a weapon similar to a quarterstaff. For non-wizard
roles it is almost exactly the same, the only difference being the
high price. All sub-roles of the wizard role can make use of one, but
only the Wizard sub-role starts with one. (They are randomly generated
but very rare - and it would be a reasonable use of a wish.)

When wielding a wizardstaff, if you are at maximum energy then you
will charge it - the stored energy increases and this is reflected in
the enchantment. 10 energy is equivalent to +1 to enchantment and
takes 50 turns to recover. The upper limit increases with level from
+1 at level 1 to +7 at level 30. (You can charge it quickly with a
scroll of enchant weapon, but not beyond its limit.)

If you attempt to cast a spell without sufficient energy, you can draw
energy from the staff to make up the difference, leaving you with 0
energy. This happens before blood magic, if you have both. The staff's
enchantment will go down again - it's always proportional to the
stored energy.

The existing artifact quartstaff 'Balance' has been made into a
wizardstaff. The + to damage has been reduced to d3 from d6 to allow
for it being easier to enchant, and the price increased.

You can exchange, unwield, drop etc. a staff without losing energy.
However if you have more than one wizardstaff, and change between them
then as you wield a staff which is not the one which you last wielded,
then it will become "attuned" to you and drain back to +0. (Otherwise
a large collection of wizardstaffs would be an almost unlimited
energy pool.)