Fire Brand is fiery
A hit against an ice vortex "melts part of" it (like the message with
water elementals, this doesn't affect damage - it's just a more
accurate message - water doesn't burn!)

A hit against a "flammable" monster instantly kills it.
These are:
All walking trees.
The wood, paper and straw golems.
And the green slime.