Remove vampires.
Remove vampires, and almost all references to them.
The drain-life attack remains (it may be used by something else).
So do vampire bats - which can bite but not transform.
This includes Alucard, potions of vampire blood, and the artifact
Vladsbane.
Potions of blood (though not vampire blood) remain, though they are
of little use.
Vlad the Impaler remains (so that the tower is still playable)
but is human.