No more leprechauns!
Re-theme: remove leprechauns.
The class is now 'lemming'. The only member is "green lemming",
which is just a placeholder (a copy of the old leprechaun).
Leprechaun halls have been replaced with lemming-houses (filled
with lemmings, no gold)
The gold-stealing behaviour remains, but behind the steal_gold()
macro which is currently not defined for any monster.