Add lemmings (class 'l', replacing leprechauns).
This required moving out the flags for troll-like reviving (highland
lemming) and gremlin-like hating light (shadow lemming), and adding
a type of summoning which is similar to lycanthropes, but doesn't
depend on night/day etc. There is a special case for desert lemmings
(which throw water bottles at you), and lemmings have various starting
items although not all may be usable yet - this has involved changing
the code to create artifacts in monster starting inventory and in the
process fixed a bug (in which the flail Unlimited Moon was being based
on a longsword).

This includes "huge groups" (if both SGROUP and LGROUP are set, then
you get more than a "large group" and the adjustment for low levels
makes less difference). It's used for the horde lemming.

The lemming-house now selects from all lemmings in level.

Lemmings will not avoid hazards - traps, water, lava - and will
actively prefer water (like eels, except that eels can swim...).