Workaround for deleting in-use objects when loading
Items were being deleted mysteriously when loading a save file.
The immediate cause was inven_inuse, so this has been changed to
just mark them as not in use and not delete them.
This will probably bring back the dreaded and infamous HUP bug, but
its infamy is limited to cheating - one more way to cheat is better
than having (as happened to me) Excalibur randomly evaporate.
Still, it is only a workaround and this needs to be investigated and
fixed properly... some day.