LOOK RARE MONSTER
Add a "rare" flag to the monster generation field.
If set, the generation frequency is treated as 10 times less,
effectively 0.1 to 0.7 - this approach was used as it allows the
existing monsters to remain unchanged.
Some monsters (mostly lemmings) are now rare (there were too many
lemmings generated because there were so many similar types).