Pirates attracted by the Pirate QA only replace weaker monsters.
The Pirate quest artifact replaced 20% of monsters with a skeletal,
damned or planar pirate. This was supposed to be a hazard, but often
wasn't as the pirates were less nasty than the typical random monster.
So now they are only spawned replacing equal or lower level monsters,
and as this reduces the overall proportion of pirates the chance is
upped to 30%, intending to produce about the same amount as before
although it will vary with level etc.