Highscore table format bug fix
It was using an old format with only one character to define a role
because the version (0.4.2) was less than the vanilla version when
it was first supported. This resulted in confusion between Priests
and Pirates for example.
It is now disabled only until 0.4.3 (the next release). It won't
take effect until a new high score table is created.